issue #115 rename : model -> model/remote/plan
