+++ client/app/src/main/java/com/example/witt/data/api/PlanService.kt
@@ -1,11 +1,11 @@
 package com.example.witt.data.api
 
-import com.example.witt.data.model.plan.get_plan.GetPlanListResponse
-import com.example.witt.data.model.plan.get_plan.GetPlanResponse
-import com.example.witt.data.model.plan.join_plan.JoinPlanResponse
-import com.example.witt.data.model.plan.join_plan.OutPlanResponse
-import com.example.witt.data.model.plan.make_plan.request.MakePlanRequest
-import com.example.witt.data.model.plan.make_plan.response.MakePlanResponse
+import com.example.witt.data.model.remote.plan.get_plan.GetPlanListResponse
+import com.example.witt.data.model.remote.plan.get_plan.GetPlanResponse
+import com.example.witt.data.model.remote.plan.join_plan.JoinPlanResponse
+import com.example.witt.data.model.remote.plan.make_plan.request.MakePlanRequest
+import com.example.witt.data.model.remote.plan.make_plan.response.MakePlanResponse
+import com.example.witt.data.model.remote.plan.out_plan.OutPlanResponse
 import retrofit2.http.*
 
 interface PlanService {
